apps/telemetry-service/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-socketio>=5.11
nats-py>=2.7
httpx>=0.27
aiokafka>=0.11
